requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==3.0.0
